fix `ZpoolOpen3::available()` when a comment is present.
Pools with a value in the `comment` field will not show up in the output from `available()`.
We observed this when testing with this library, because pools created with this library
and then subsequently exported have an empty comment.

Looking at the [ZFS source](https://github.com/openzfs/zfs/blob/baca06c258e07522165cb8e33ff2c0224ad0da57/cmd/zpool/zpool_main.c#L3053)
it looks like the `comment:` line comes after the `action:` line.

In testing this I wanted to determine if comment should be a multi-line string or just text.
I tried creating and exporting a pool with a `\n` in the comment field and found that the
`\n` was reproduced in its escaped form.